Refactor by move packages
