+++ pom.xml
@@ -48,12 +48,6 @@
 
 		<dependency>
 			<groupId>de.uni.bielefeld.sc.hterhors.psink.obie</groupId>
-			<artifactId>SoccerPlayerOntology</artifactId>
-			<version>0.0.1-SNAPSHOT</version>
-		</dependency>
-
-		<dependency>
-			<groupId>de.uni.bielefeld.sc.hterhors.psink.obie</groupId>
 			<artifactId>OWL2JavaBin</artifactId>
 			<version>0.0.1-SNAPSHOT</version>
 		</dependency>
