+++ src/main/java/de/hterhors/obie/projects/soccerplayer/tools/CreateSoccerPlayerOntologyJavaBinaries.java
@@ -1,0 +1,18 @@
+package de.hterhors.obie.projects.soccerplayer.tools;
+
+import de.hterhors.obie.projects.soccerplayer.ie.SoccerPlayerOntologyEnvironment;
+import de.uni.bielefeld.sc.hterhors.psink.obie.ontology.owl2javabin.OWLToJavaBinaries;
+
+public class CreateSoccerPlayerOntologyJavaBinaries {
+
+	public static void main(String[] args) throws Exception {
+
+		OWLToJavaBinaries builder = new OWLToJavaBinaries(SoccerPlayerOntologyEnvironment.getInstance());
+
+		builder.buildAndWriteClasses();
+
+		builder.buildAndWriteInterfaces();
+
+	}
+
+}
