--- src/main/java/de/uni/bielefeld/sc/hterhors/psink/projects/soccerplayer/ie/ner/regex/SoccerPlayerRegExNEL.java
@@ -1,140 +1,0 @@
-package de.uni.bielefeld.sc.hterhors.psink.projects.soccerplayer.ie.ner.regex;
-
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Set;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
-import de.uni.bielefeld.sc.hterhors.psink.obie.core.OntologyAnalyzer;
-import de.uni.bielefeld.sc.hterhors.psink.obie.core.ontology.AbstractOBIEIndividual;
-import de.uni.bielefeld.sc.hterhors.psink.obie.core.ontology.OntologyInitializer;
-import de.uni.bielefeld.sc.hterhors.psink.obie.core.ontology.annotations.DatatypeProperty;
-import de.uni.bielefeld.sc.hterhors.psink.obie.core.ontology.interfaces.IOBIEThing;
-import de.uni.bielefeld.sc.hterhors.psink.obie.ie.dtinterpreter.IDatatypeInterpretation;
-import de.uni.bielefeld.sc.hterhors.psink.obie.ie.ner.regex.AbstractRegExNER;
-import de.uni.bielefeld.sc.hterhors.psink.obie.ie.ner.regex.BasicRegExPattern;
-import de.uni.bielefeld.sc.hterhors.psink.obie.projects.soccerplayer.ontology.interfaces.ISoccerPlayer;
-import de.uni.bielefeld.sc.hterhors.psink.obie.projects.soccerplayer.ontology.interfaces.ISoccerPlayerThing;
-import de.uni.bielefeld.sc.hterhors.psink.projects.soccerplayer.ie.SoccerPlayerOntologyEnvironment;
-import de.uni.bielefeld.sc.hterhors.psink.projects.soccerplayer.ie.dtinterpreter.SoccerPlayerInterpreterProvider;
-
-public class SoccerPlayerRegExNEL extends AbstractRegExNER<ISoccerPlayerThing> {
-
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
-	private static final int MIN_NER_LENGTH = 2;
-
-	public static Logger log = LogManager.getFormatterLogger(SoccerPlayerRegExNEL.class.getSimpleName());
-
-	public SoccerPlayerRegExNEL(Set<Class<? extends IOBIEThing>> rootClasses) {
-		super(rootClasses);
-	}
-
-	protected Map<Class<? extends ISoccerPlayerThing>, Set<Pattern>> generateHandMadeCrossReferencesForClasses(
-			Map<Class<? extends ISoccerPlayerThing>, Set<Pattern>> regularExpressionPattern,
-			Class<? extends IOBIEThing> rootClassType) {
-		/**
-		 * Add cross reference pattern. A cross reference pattern is used for extend
-		 * specific fields by pattern of other fields.
-		 */
-		return Collections.emptyMap();
-	}
-
-	protected Map<Class<? extends ISoccerPlayerThing>, Set<Pattern>> generateCrossReferencePatternForClasses(
-			Class<? extends IOBIEThing> rootClassType) {
-		return Collections.emptyMap();
-	}
-
-	protected Map<Class<? extends ISoccerPlayerThing>, Set<Pattern>> addHandMadePatternForClasses(
-			Class<? extends IOBIEThing> rootClassType) {
-
-		Map<Class<? extends ISoccerPlayerThing>, Set<Pattern>> relatedHandMadepattern = new HashMap<>();
-
-		Map<Class<? extends ISoccerPlayerThing>, Set<Pattern>> allHandMadepattern = SoccerPlayerRegExPattern
-				.getHandMadePattern();
-
-		for (Class<? extends IOBIEThing> classType : OntologyAnalyzer.getRelatedClassTypesUnderRoot(rootClassType)) {
-
-			if (classType.isAnnotationPresent(DatatypeProperty.class) && !allHandMadepattern.containsKey(classType)) {
-				log.warn("WARN!!! No hand made pattern for data type property: " + classType.getSimpleName());
-			}
-
-			if (allHandMadepattern.containsKey(classType)) {
-				relatedHandMadepattern.put((Class<ISoccerPlayerThing>) classType, allHandMadepattern.get(classType));
-			}
-		}
-
-		if (rootClassType == ISoccerPlayer.class) {
-		} else {
-			/*
-			 * TODO: add more hand made pattern HandMade Pattern
-			 */
-			log.warn("STOP PROCESS! Unknown class for handmade pattern: " + rootClassType.getSimpleName());
-			OntologyAnalyzer.debug = true;
-			OntologyAnalyzer.getRelatedClassTypesUnderRoot(rootClassType);
-			throw new IllegalArgumentException(
-					"STOP PROCESS! Unknown class for handmade pattern: " + rootClassType.getSimpleName());
-		}
-		return relatedHandMadepattern;
-	}
-
-	@Override
-	protected Map<Class<? extends ISoccerPlayerThing>, Set<Pattern>> addFurtherPatternForClasses() {
-		return Collections.emptyMap();
-	}
-
-	@Override
-	protected Map<Class<? extends ISoccerPlayerThing>, Set<Pattern>> addPlainRegExPatternForClasses(
-			Class<? extends IOBIEThing> interfaceClassType) {
-		return new SoccerPlayerRegExPattern().autoGeneratePatternForClasses(interfaceClassType);
-	}
-
-	@Override
-	protected IDatatypeInterpretation getSemanticInterpretation(Class<? extends ISoccerPlayerThing> dataTypeClass,
-			Matcher matcher) {
-		return SoccerPlayerInterpreterProvider.getInstance().interpret(dataTypeClass, matcher);
-	}
-
-	@Override
-	protected int getMinNERLength() {
-		return MIN_NER_LENGTH;
-	}
-
-	@Override
-	protected Map<AbstractOBIEIndividual, Set<Pattern>> addHandMadePatternForIndividuals(
-			Class<? extends IOBIEThing> rootClassType) {
-		return Collections.emptyMap();
-	}
-
-	@Override
-	protected Map<AbstractOBIEIndividual, Set<Pattern>> addFurtherPatternForIndividuals() {
-		return Collections.emptyMap();
-	}
-
-	@Override
-	protected Map<AbstractOBIEIndividual, Set<Pattern>> generateHandMadeCrossReferencesForIndividuals(
-			Map<AbstractOBIEIndividual, Set<Pattern>> regularExpressionPattern,
-			Class<? extends IOBIEThing> rootClassType) {
-		return Collections.emptyMap();
-	}
-
-	@Override
-	protected Map<AbstractOBIEIndividual, Set<Pattern>> generateCrossReferencePatternForIndividuals(
-			Class<? extends IOBIEThing> rootClassType) {
-		return Collections.emptyMap();
-	}
-
-	@Override
-	protected Map<AbstractOBIEIndividual, Set<Pattern>> addPlainRegExPatternForIndividuals(
-			Class<? extends IOBIEThing> rootClassType) {
-		return new SoccerPlayerRegExPattern().autoGeneratePatternForIndividuals(rootClassType);
-	}
-
-}
